First steps
Adding ZF context and Dialog to get away from old ESX menus! FiveM is all i use so that is only why it is in there! i will be adding back the options to use Qtarget or BT-Target! they will be hashed out and require you to unhash them and hash out the FiveM target stuff!!! super easy!